refactor(plugins): move test files to /tests folder
