+++ public/app/features/plugins/tests/AppRootPage.test.tsx
@@ -1,9 +1,9 @@
 import { act, render, screen } from '@testing-library/react';
 import React, { Component } from 'react';
-import AppRootPage from './AppRootPage';
-import { getPluginSettings } from './PluginSettingsCache';
-import { importAppPlugin } from './plugin_loader';
-import { getMockPlugin } from './__mocks__/pluginMocks';
+import AppRootPage from '../AppRootPage';
+import { getPluginSettings } from '../PluginSettingsCache';
+import { importAppPlugin } from '../plugin_loader';
+import { getMockPlugin } from '../__mocks__/pluginMocks';
 import { AppPlugin, PluginType, AppRootProps, NavModelItem } from '@grafana/data';
 import { Route, Router } from 'react-router-dom';
 import { locationService, setEchoSrv } from '@grafana/runtime';
